refactor(#3038): move turn bridge analytics persistence
